non matrix outlet added
matrix outlets can be added by the standard min outlet<> decleration, and the correct count of matrix outlets must be provided to the mop constructor in our jitclass_setup override method.

but to add non-matrix outlets we must override mop_setup and manage the outlet in our class.

we also must provide a custom assist handler.